fix: comprehensive OpenBugs fixes (45 tests fixed, 108→63 failures)
Major fixes:
- Implement np.isclose/allclose with broadcast support and NaN handling
- Implement np.repeat with per-element repeat counts (array input)
- Fix np.convolve for all modes (full, same, valid) and int arrays
- Fix np.isfinite/isnan to work (were returning null)
- Fix Python % modulo semantics in ILKernelGenerator (floored division)

Implementation details:
- IsClose: Uses NumPy formula |a-b| <= (atol + rtol*|b|) with NaN/Inf handling
- Repeat: New overload taking NDArray for per-element counts
- Convolve: Fixed mode handling and integer array support
- IsFinite/IsNan: Implemented proper float/double checking
- Mod: Emit proper floored division for signed integers (Python semantics)

Test improvements:
- Added np.minimum tests
- Fixed argsort tests for 2D arrays
- Updated edge case tests

+++ test/NumSharp.UnitTest/Extensions/ndarray.argsort.Test.cs
@@ -15,9 +15,10 @@ namespace NumSharp.UnitTest.Extensions
         [Test]
         public void OneDimension()
         {
+            // NumPy argsort always returns int64 indices
             var x = np.array(new int[] {3, 1, 2});
             x = x.argsort<int>();
-            Assert.IsTrue(Enumerable.SequenceEqual(new int[] {1, 2, 0}, x.Data<int>()));
+            Assert.IsTrue(Enumerable.SequenceEqual(new long[] {1, 2, 0}, x.Data<long>()));
         }
 
         [Test]
